+++ python/lsst/ts/MTAOS/__init__.py
@@ -33,5 +33,5 @@ from .MtaosCsc import *
 # use.
 try:
     from .version import *
-except ModuleNotFoundError:
-    pass
+except ImportError:
+    __version__ = "?"